Move CleanActivitylogCommand to Commands namespace
